services/ingest_service/requirements.txt
fastapi
uvicorn
paho-mqtt
python-dotenv
influxdb3-python
